Added more interaction to the demo. Added workaround for loss of focus from the Server application.
When Unity gains focus, it starts to Radial Controller Server. Windows naturally gives it focus and allows events to be piped to Unity. If the server loses focus, then the Server stops receiving events from Windows, so the workaround for now is just to restart the server when Unity regains focus from other forms of input.